backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "my project"
authors = ["Your Name <email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
fastapi = "0.128.0"
uvicorn = ">=0.40.0,<0.41.0"
sqlalchemy = ">=2.0.46,<3.0.0"
aiogram = ">=3.24.0,<4.0.0"
pydantic-settings = ">=2.12.0,<3.0.0"
python-multipart = ">=0.0.21,<0.0.22"
aiosqlite = ">=0.22.1,<0.23.0"
colorlog = ">=6.10.1,<7.0.0"
sqladmin = ">=0.22.0,<0.23.0"
email-validator = ">=2.3.0,<3.0.0"
itsdangerous = ">=2.2.0,<3.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
bcrypt = ">=4.0.0,<5.0.0"
alembic = "^1.18.1"
starlette = ">=0.40.0,<0.51.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
